Create optpath.sh
Initial commit, to be used with configurations that have installed applications in the /opt hierarchy. Note that this makdes no assumption of the location of related files, which may be located in /opt/name.tld/etc or /etc/opt/name.tld. We recommend using domain name-based namespaces for directory names.